+++ src/memory/allocation/resource_types/immutable_resource.hpp
@@ -93,8 +93,7 @@ struct __immutable_memory_resource : public __core_memory_resource<T> {
   {
     if ( len == 0 ) [[unlikely]]
       return;
-    if ( __core_memory_resource<T>::alive() )
-      Alloc::destroy(__core_memory_resource<T>::operator*());
+    if ( __core_memory_resource<T>::alive() ) Alloc::destroy(__core_memory_resource<T>::operator*());
     __core_memory_resource<T>::accept(Alloc::create(len * (sizeof(T) / sizeof(byte))));
   }
